working to try and add modal update function; added separate file for modal but not working
